requirements.txt
streamlit==1.53.0
google-genai==1.59.0